nativedsl: pool macro params and for-bindings as ChildRange
Convert MacroConfig.params and ForConfig.bindings from Vec<Param> to a
ChildRange into a new shared AstPools.params pool, removing one heap
allocation per macro and per for-loop. push_params returns ChildRange
directly (not Option): both counts are already bounded to u8::MAX at parse
time (their indices are u8), so the length always fits ChildRange::len.

Pure representation change; no behavior change.

+++ crates/generate/src/nativedsl/expand_macro_calls.rs
@@ -126,7 +126,7 @@ fn expand_one_call(
     // Snapshot before recursive calls reborrow shared mutably.
     let config = shared.pools.get_macro(macro_id);
     let kind = config.kind;
-    let param_count = config.params.len();
+    let param_count = config.params.len as usize;
     let body_id = config.body;
     let MacroKind::RuleSet = kind else {
         return Err(ExpandError::new(
